refactor: getOrCreateCategory method 리팩토링 - #149
getOrCreateCategory method를
custom repository로 옮기고 코드를 정리함
