+++ user_logs/1.txt
@@ -1,3 +1,5 @@
-{"categoryId": 16,"savedAt": 1674218419211}
-{"categoryId": 16,"savedAt": 1674218423926}
-{"categoryId": 13,"savedAt": 1674218427094}
+{"categoryId": 13,"savedAt": 1674228254021}
+{"categoryId": 13,"savedAt": 1674228257056}
+{"categoryId": 16,"savedAt": 1674228317271}
+{"categoryId": 16,"savedAt": 1674646370377}
+{"categoryId": 14,"savedAt": 1674646406408}
